Make sure we have a local file for reading rasters.
