+++ geodata/models/raster/reader.py
@@ -3,6 +3,7 @@ from celery.utils.log import get_task_logger
 from django.contrib.gis.gdal import SpatialReference
 from django.contrib.gis.geos import Polygon
 import rasterio
+from rgd.utility import _field_file_to_local_path
 
 # from django.core.files.base import ContentFile
 # import io
@@ -41,82 +42,82 @@ class RasterEntryReader(_ReaderRoutine):
         """
         # Fetch the raster file this Layer corresponds to
         self.rfe = RasterFile.objects.get(id=self.model_id)
-        file_path = self.rfe.raster_file.name
+        with _field_file_to_local_path(self.rfe.raster_file) as file_path:
 
-        logger.info(f'The raster file path: {file_path}')
+            logger.info(f'The raster file path: {file_path}')
 
-        raster_querry = RasterEntry.objects.filter(raster_file=self.rfe)
-        if len(raster_querry) < 1:
-            self.raster_entry = RasterEntry()
-        elif len(raster_querry) == 1:
-            self.raster_entry = raster_querry.first()
-            # Clear out associated entries because they could be invalid
-            BandMetaEntry.objects.filter(parent_raster=self.raster_entry).delete()
-            ConvertedRasterFile.objects.filter(source_raster=self.raster_entry).delete()
-        else:
-            # This should never happen because it is a foreign key
-            raise RuntimeError('multiple raster entries found for this file.')
+            raster_querry = RasterEntry.objects.filter(raster_file=self.rfe)
+            if len(raster_querry) < 1:
+                self.raster_entry = RasterEntry()
+            elif len(raster_querry) == 1:
+                self.raster_entry = raster_querry.first()
+                # Clear out associated entries because they could be invalid
+                BandMetaEntry.objects.filter(parent_raster=self.raster_entry).delete()
+                ConvertedRasterFile.objects.filter(source_raster=self.raster_entry).delete()
+            else:
+                # This should never happen because it is a foreign key
+                raise RuntimeError('multiple raster entries found for this file.')
 
-        self.raster_entry.raster_file = self.rfe
+            self.raster_entry.raster_file = self.rfe
 
-        with rasterio.open(file_path) as src:
-            self.raster_entry.number_of_bands = src.count
-            self.raster_entry.driver = src.driver
-            self.raster_entry.crs = src.crs.to_proj4()
-            self.raster_entry.origin = [src.bounds.left, src.bounds.top]
-            self.raster_entry.extent = [
-                src.bounds.left,
-                src.bounds.bottom,
-                src.bounds.right,
-                src.bounds.top,
-            ]
-            self.raster_entry.resolution = src.res
-            self.raster_entry.height = src.shape[0]
-            self.raster_entry.width = src.shape[1]
-            self.raster_entry.transform = src.transform.to_gdal()  # TODO: check this
-            # A catch-all metadata feild:
-            # TODO: self.raster_entry.metadata =
+            with rasterio.open(file_path) as src:
+                self.raster_entry.number_of_bands = src.count
+                self.raster_entry.driver = src.driver
+                self.raster_entry.crs = src.crs.to_proj4()
+                self.raster_entry.origin = [src.bounds.left, src.bounds.top]
+                self.raster_entry.extent = [
+                    src.bounds.left,
+                    src.bounds.bottom,
+                    src.bounds.right,
+                    src.bounds.top,
+                ]
+                self.raster_entry.resolution = src.res
+                self.raster_entry.height = src.shape[0]
+                self.raster_entry.width = src.shape[1]
+                self.raster_entry.transform = src.transform.to_gdal()  # TODO: check this
+                # A catch-all metadata feild:
+                # TODO: self.raster_entry.metadata =
 
-            # thumbnail = RasterEntryReader.create_thumbnail(src)
-            # self.raster_entry.thumbnail.save('thumbnail.jpg', thumbnail, save=True)
+                # thumbnail = RasterEntryReader.create_thumbnail(src)
+                # self.raster_entry.thumbnail.save('thumbnail.jpg', thumbnail, save=True)
 
-            coords = (
-                (src.bounds.left, src.bounds.top),
-                (src.bounds.right, src.bounds.top),
-                (src.bounds.right, src.bounds.bottom),
-                (src.bounds.left, src.bounds.bottom),
-                (src.bounds.left, src.bounds.top),  # Close the loop
-            )
+                coords = (
+                    (src.bounds.left, src.bounds.top),
+                    (src.bounds.right, src.bounds.top),
+                    (src.bounds.right, src.bounds.bottom),
+                    (src.bounds.left, src.bounds.bottom),
+                    (src.bounds.left, src.bounds.top),  # Close the loop
+                )
 
-            spatial_ref = SpatialReference(src.crs.to_wkt())
-            logger.info(f'Raster footprint SRID: {spatial_ref.srid}')
-            # This will convert the Polygon to the DB's SRID
-            self.raster_entry.footprint = Polygon(coords, srid=spatial_ref.srid)
+                spatial_ref = SpatialReference(src.crs.to_wkt())
+                logger.info(f'Raster footprint SRID: {spatial_ref.srid}')
+                # This will convert the Polygon to the DB's SRID
+                self.raster_entry.footprint = Polygon(coords, srid=spatial_ref.srid)
 
-            # TODO: now create a `BandMetaEntry` entry for each band
-            #       this will collect statistics on the bands
-            self.band_entries = []
-            for i in range(src.count):
-                band_meta = BandMetaEntry()
-                band_meta.parent_raster = self.raster_entry
-                try:
-                    band_meta.description = src.descriptions[i]
-                except IndexError:
-                    pass
-                try:
-                    band_meta.nodata_value = src.nodatavals[i]
-                except IndexError:
-                    pass
-                try:
-                    band_meta.dtype = src.dtypes[i]
-                except IndexError:
-                    pass
-                # TODO: Compute statistics: min, max, mean, std
-                #       We'd have to load the bands into memroy...
-                #       might not be good.
+                # TODO: now create a `BandMetaEntry` entry for each band
+                #       this will collect statistics on the bands
+                self.band_entries = []
+                for i in range(src.count):
+                    band_meta = BandMetaEntry()
+                    band_meta.parent_raster = self.raster_entry
+                    try:
+                        band_meta.description = src.descriptions[i]
+                    except IndexError:
+                        pass
+                    try:
+                        band_meta.nodata_value = src.nodatavals[i]
+                    except IndexError:
+                        pass
+                    try:
+                        band_meta.dtype = src.dtypes[i]
+                    except IndexError:
+                        pass
+                    # TODO: Compute statistics: min, max, mean, std
+                    #       We'd have to load the bands into memroy...
+                    #       might not be good.
 
-                # Keep track
-                self.band_entries.append(band_meta)
+                    # Keep track
+                    self.band_entries.append(band_meta)
 
         return True
