+++ rgd/__init__.py
@@ -1,3 +1,4 @@
 from .celery import app as celery_app
+from . import utility
 
-__all__ = ('celery_app',)
+__all__ = ('celery_app', 'utility')
